requirements.txt

numpy
torch 
torchvision 
Pillow==9.4.0
opencv-python>=4.1.0
matplotlib
imutils
dlib
PyQt5